requirements.txt
google-adk==1.19.0
google-cloud-aiplatform[agent_engines,evaluation,adk]
pandas
mcp[fastmcp]
uvicorn[standard]
google-cloud-bigquery
db-dtypes
tabulate

